rename smuggler modules to *_converter for clarity
