+++ pyrit/prompt_converter/token_smuggling/__init__.py
@@ -1,12 +1,12 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from pyrit.prompt_converter.token_smuggling.ascii_smuggler import AsciiSmugglerConverter
-from pyrit.prompt_converter.token_smuggling.sneaky_bits_smuggler import SneakyBitsSmuggler
-from pyrit.prompt_converter.token_smuggling.variation_selector_smuggler import VariationSelectorSmuggler
+from pyrit.prompt_converter.token_smuggling.ascii_smuggler_converter import AsciiSmugglerConverter
+from pyrit.prompt_converter.token_smuggling.sneaky_bits_smuggler_converter import SneakyBitsSmugglerConverter
+from pyrit.prompt_converter.token_smuggling.variation_selector_smuggler_converter import VariationSelectorSmugglerConverter
 
 __all__ = [
     "AsciiSmugglerConverter",
-    "SneakyBitsSmuggler",
-    "VariationSelectorSmuggler",
+    "SneakyBitsSmugglerConverter",
+    "VariationSelectorSmugglerConverter",
 ]
