+++ pyrit/prompt_converter/token_smuggling/variation_selector_smuggler_converter.py
@@ -9,7 +9,7 @@ from pyrit.prompt_converter.token_smuggling.base import SmugglerConverter
 logger = logging.getLogger(__name__)
 
 
-class VariationSelectorSmuggler(SmugglerConverter):
+class VariationSelectorSmugglerConverter(SmugglerConverter):
     """
     Encodes and decodes text using Unicode Variation Selectors.
